mcpwm.c: Use only FORWARDS comm step table; walk it in reverse to reverse.
This is a major change that simplifies the code a lot. Use the fact that
it's possible to simply walk the commutation table in reverse to reverse
the motor, rather than using another table with phase 2 & 3 switched.

Make the direction variable take the values 1, -1 and take 'direction'
steps at a time when running sensorless, and do a simple conversion
between a forward step and the reverse step corresponding to the same
rotor position, in the same table, for hall sensor commutation in reverse.

Also, using only one table (where a certain step always corresponds to
a certain neg/off/pos bridge config) means it's possible to remove a lot
of "direction" dependent code.

See the comment starting at line 1953 in mcpwm.c for more detailed info.